Move excel-specific decorators into types.ts
